requirements.txt
aerich==0.8.2 ; python_full_version < '3.9'
aerich==0.9.2 ; python_full_version >= '3.9'
aiosqlite==0.20.0 ; python_full_version < '3.9'
aiosqlite==0.21.0 ; python_full_version >= '3.9'
annotated-types==0.7.0 ; python_full_version < '3.9'
anyio==4.5.2 ; python_full_version < '3.9'
anyio==4.11.0 ; python_full_version >= '3.9'
asyncclick==8.1.7.2 ; python_full_version < '3.9'
asyncclick==8.1.8 ; python_full_version == '3.9.*'
asyncclick==8.3.0.3 ; python_full_version == '3.10.*'
asyncclick==8.3.0.7 ; python_full_version >= '3.11'
click==8.1.8 ; python_full_version < '3.10'
click==8.3.0 ; python_full_version >= '3.10'
colorama==0.4.6 ; sys_platform == 'win32'
dictdiffer==0.9.0
dynaconf==3.2.12
eval-type-backport==0.2.2 ; python_full_version < '3.9'
exceptiongroup==1.3.0 ; python_full_version < '3.11'
idna==3.11
iso8601==2.1.0 ; python_full_version < '4'
orjson==3.10.15 ; python_full_version < '3.9'
orjson==3.11.4 ; python_full_version >= '3.9'
pyaes==1.6.1
pydantic==2.10.6 ; python_full_version < '3.9'
pydantic-core==2.27.2 ; python_full_version < '3.9'
pypika-tortoise==0.3.2 ; python_full_version < '3.9'
pypika-tortoise==0.6.2 ; python_full_version >= '3.9' and python_full_version < '4'
pyrogram==2.0.106
pysocks==1.7.1
python-dotenv==1.0.1 ; python_full_version < '3.9'
python-dotenv==1.1.1 ; python_full_version >= '3.9'
pytz==2025.2
sniffio==1.3.1
structlog==25.5.0
tgcrypto==1.2.5
tomli==2.3.0
tortoise-orm==0.23.0 ; python_full_version < '3.9'
tortoise-orm==0.25.1 ; python_full_version >= '3.9'
typing-extensions==4.13.2 ; python_full_version < '3.9'
typing-extensions==4.15.0 ; python_full_version >= '3.9'
